feat(backend): specify method for calculating regime switch time.
Adds `method` argument to `get_switch_time`. The user can specify which
method is used to calculate the local time of changes between stability
conditions. This argument may be deprecated for a more automated
solution.

Updates methods in the metrics/calculations module to reflect changes
in how data is passed.

Refs: ST-3, ST-6, ST-40, ST-97, ST-98